Add data migration to new Usuari custom user model
There's a new data migration that uses a couple of SQL queries to
copy the old auth_user and auth_user_groups table's data to the
new tables for the Usuari model.

With the new migration, the temp SQL file with this query isn't
needed.

Also fixed an error in the Usuari model.