Esdeveniments (Events)
Add events to the web-site. An event can be a one off or can
be part of a series of events.

The publication and seminar archive pages have been removed
and replaced with a calendar, which shows future events and
seminars and an archive page which contains a list of past
seminars and events as well as listing all the publications.

Translations for Castilian and Catalan are included.